Fix high Regressor Model Score
The regressor model was giving a very high score
because I was trying to predict the exact closing
price of the next day, these prices do not vary
much from day to day in the index like sensex, hence
any guess would be very close to the actual value.
To fix this I changed the target variable to be the
percentage change in closing price the next day
compared to the current day. Now I am getting more
realistic scores from the model.